♻️ E1BG2Q doctor: split hook readiness hotspot
