+++ packages/agentplane/src/commands/doctor/hook-readiness.ts
@@ -1,0 +1,214 @@
+import fs from "node:fs/promises";
+import path from "node:path";
+
+import { renderDiagnosticFinding } from "../shared/diagnostics.js";
+import {
+  fileIsManaged,
+  HOOK_MARKER,
+  HOOK_NAMES,
+  resolveGitHooksDir,
+  SHIM_MARKER,
+} from "../hooks/shared.js";
+
+async function pathExists(absPath: string): Promise<boolean> {
+  try {
+    await fs.access(absPath);
+    return true;
+  } catch {
+    return false;
+  }
+}
+
+async function readTextIfExists(absPath: string): Promise<string | null> {
+  try {
+    return await fs.readFile(absPath, "utf8");
+  } catch {
+    return null;
+  }
+}
+
+function unquoteShellSingleQuoted(value: string): string {
+  return value.replaceAll(String.raw`'\''`, "'");
+}
+
+function extractInstallBinFromShim(shimText: string): string | null {
+  const match = shimText.match(/\nINSTALL_BIN='((?:'\\''|[^'])*)'/);
+  if (!match) return null;
+  return unquoteShellSingleQuoted(match[1] ?? "").trim() || null;
+}
+
+async function resolveManagedHookNames(repoRoot: string): Promise<{
+  hooksDir: string | null;
+  managedHooks: string[];
+  unmanagedHooks: string[];
+}> {
+  let hooksDir: string;
+  try {
+    hooksDir = await resolveGitHooksDir(repoRoot);
+  } catch {
+    return { hooksDir: null, managedHooks: [], unmanagedHooks: [] };
+  }
+
+  const managedHooks: string[] = [];
+  const unmanagedHooks: string[] = [];
+  for (const hook of HOOK_NAMES) {
+    const hookPath = path.join(hooksDir, hook);
+    if (!(await pathExists(hookPath))) continue;
+    if (await fileIsManaged(hookPath, HOOK_MARKER)) managedHooks.push(hook);
+    else unmanagedHooks.push(hook);
+  }
+  return { hooksDir, managedHooks, unmanagedHooks };
+}
+
+function renderUnmanagedHooksFinding(unmanagedHooks: string[]): string {
+  return renderDiagnosticFinding({
+    severity: "INFO",
+    state: "unmanaged git hooks are present and will not be overwritten by AgentPlane",
+    likelyCause: "the repository has custom hook files outside the AgentPlane managed hook set",
+    nextAction: {
+      command: "agentplane hooks install",
+      reason:
+        "refresh managed hooks only after manually reconciling custom hook files that should remain custom",
+    },
+    details: [`Unmanaged hooks: ${unmanagedHooks.join(", ")}`],
+  });
+}
+
+export async function checkManagedHookShimReadiness(repoRoot: string): Promise<string[]> {
+  const findings: string[] = [];
+  const { managedHooks, unmanagedHooks } = await resolveManagedHookNames(repoRoot);
+  const shimRelPath = ".agentplane/bin/agentplane";
+  const shimPath = path.join(repoRoot, shimRelPath);
+  const shimText = await readTextIfExists(shimPath);
+  const managedShim = shimText?.includes(SHIM_MARKER) === true;
+
+  if (managedHooks.length === 0 && unmanagedHooks.length === 0 && !shimText) return findings;
+
+  if (managedHooks.length > 0 && !shimText) {
+    findings.push(
+      renderDiagnosticFinding({
+        severity: "WARN",
+        state: "managed git hooks are installed but the AgentPlane hook shim is missing",
+        likelyCause:
+          "the repository hook files still call .agentplane/bin/agentplane, but that shim was deleted or never committed",
+        nextAction: {
+          command: "agentplane hooks install",
+          reason: "recreate the managed hook shim and refresh hook files",
+        },
+        details: [`Managed hooks: ${managedHooks.join(", ")}`],
+      }),
+    );
+    return findings;
+  }
+
+  if (shimText && !managedShim) {
+    findings.push(
+      renderDiagnosticFinding({
+        severity: "WARN",
+        state: "AgentPlane hook shim path exists but is not managed by AgentPlane",
+        likelyCause:
+          "a custom .agentplane/bin/agentplane file is shadowing the managed hook runner shim",
+        nextAction: {
+          command:
+            "inspect .agentplane/bin/agentplane, remove the custom file if unintended, then run agentplane hooks install",
+          reason: "restore the managed shim without overwriting custom executable content silently",
+        },
+        details: [`Shim path: ${shimRelPath}`],
+      }),
+    );
+    return findings;
+  }
+
+  if (!shimText) {
+    if (unmanagedHooks.length > 0) {
+      findings.push(renderUnmanagedHooksFinding(unmanagedHooks));
+    }
+    return findings;
+  }
+
+  const installedBin = extractInstallBinFromShim(shimText);
+  if (!installedBin) {
+    findings.push(
+      renderDiagnosticFinding({
+        severity: "WARN",
+        state:
+          "managed AgentPlane hook shim uses a stale format without an installed runner fallback",
+        likelyCause:
+          "the shim was generated by an older AgentPlane release before installed-binary fallback was embedded",
+        nextAction: {
+          command: "agentplane hooks install",
+          reason: "rewrite the managed shim with the current installed-runtime fallback contract",
+        },
+        details: [`Shim path: ${shimRelPath}`],
+      }),
+    );
+  } else if (!(await pathExists(installedBin))) {
+    findings.push(
+      renderDiagnosticFinding({
+        severity: "WARN",
+        state: "managed AgentPlane hook shim points to a missing installed runner",
+        likelyCause:
+          "AgentPlane was reinstalled, removed, or moved after hooks were installed in this repository",
+        nextAction: {
+          command: "agentplane hooks install",
+          reason:
+            "refresh the embedded installed runner path after reinstalling or updating AgentPlane",
+        },
+        details: [`Embedded runner: ${installedBin}`, `Shim path: ${shimRelPath}`],
+      }),
+    );
+  }
+
+  const hasEnvFallback = shimText.includes("AGENTPLANE_HOOK_RUNNER");
+  const hasNpxFallback = shimText.includes("AGENTPLANE_HOOK_ALLOW_NPX");
+  if (!hasEnvFallback || !hasNpxFallback) {
+    findings.push(
+      renderDiagnosticFinding({
+        severity: "WARN",
+        state: "managed AgentPlane hook shim is missing current fallback branches",
+        likelyCause:
+          "the repository still has an older managed shim that cannot use env-runner or explicit npx fallback recovery",
+        nextAction: {
+          command: "agentplane hooks install",
+          reason: "rewrite the shim with the current fallback chain",
+        },
+        details: [
+          `Shim path: ${shimRelPath}`,
+          `AGENTPLANE_HOOK_RUNNER fallback: ${hasEnvFallback ? "present" : "missing"}`,
+          `AGENTPLANE_HOOK_ALLOW_NPX fallback: ${hasNpxFallback ? "present" : "missing"}`,
+        ],
+      }),
+    );
+  }
+
+  if (managedHooks.includes("pre-push")) {
+    const repoPrePushScript = path.join(repoRoot, "scripts", "run-pre-push-hook.mjs");
+    const packageJsonText = await readTextIfExists(path.join(repoRoot, "package.json"));
+    if (!(await pathExists(repoPrePushScript)) && packageJsonText) {
+      findings.push(
+        renderDiagnosticFinding({
+          severity: "INFO",
+          state: "managed pre-push hook will use installed clean-project fallback checks",
+          likelyCause:
+            "this project does not ship a repository-local scripts/run-pre-push-hook.mjs",
+          nextAction: {
+            command:
+              "add project-local package scripts such as format:check or test, or keep the installed fallback",
+            reason:
+              "make pre-push behavior explicit when the project needs stricter repository-specific checks",
+          },
+          details: [
+            "Missing repository script: scripts/run-pre-push-hook.mjs",
+            "Installed fallback skips package scripts that are not declared by the clean project.",
+          ],
+        }),
+      );
+    }
+  }
+
+  if (unmanagedHooks.length > 0) {
+    findings.push(renderUnmanagedHooksFinding(unmanagedHooks));
+  }
+
+  return findings;
+}
